Output Single Trial Onsets
When working with neural data, sometimes we want to examine at the trial-level (e.g., I want to be able to correlate behavior with neural activity across multiple trials). I've added an argument that, when true, will output a separate onset file for each trial or observation within the experiment.